Task #4585: fix accidentally broken tBeamFormerKernel: limit by channels per sb. Also turn Kernel's preproc values from size_t into unsigned.

+++ branches/Cobalt-Task4585/RTCP/Cobalt/GPUProc/src/cuda/Kernels/Kernel.h
@@ -45,11 +45,11 @@ namespace LOFAR
       struct Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrStations;
-        size_t nrChannelsPerSubband;
-        size_t nrSamplesPerChannel;
-        size_t nrSamplesPerSubband;
-        size_t nrPolarizations;
+        unsigned nrStations;
+        unsigned nrChannelsPerSubband;
+        unsigned nrSamplesPerChannel;
+        unsigned nrSamplesPerSubband;
+        unsigned nrPolarizations;
         bool dumpBuffers;
         std::string dumpFilePattern;
       };
